Konflux onboarding: introduce a generic artifacts lockfile
This commit introduces a "artifacts.lock.yaml" file which will be used
by a future commit in order to fetch the Go and NodeJs binaries directly
from their official registries.

This is due to a change we will need in comparison to our current
Dockerfile: instead of fetching whole images from docker.io to grab the
binaries, we will restrict what we fetch to specific files, making our
process more aligned with the higher requirements that Konflux has.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>