Konflux onboarding: introduce the rpms.lock.yaml file
In order to enable the prefetch and offline install of RPMs, we need to
generate a lockfile using the rpms-lockfile-prototype¹ project.

The ubi.repo file was copied straight from a UBI9 image, and is needed
by the rpms-lockfile-prototype in order to generate the final lockfile.

The newly added files will be leveraged in a future commit that will
introduce a Dockerfile dedicated to Konflux.

¹ https://github.com/konflux-ci/rpm-lockfile-prototype

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>