Konflux onboarding: add a Konflux dedicated Dockerfile
The Konflux Dockerfile is derived from the current Dockerfile present in
the root of the repository, and has the necessary modifications needed
to allow it to be built hermetically with prefetched dependencies. It
leverages the previously rpms and generic artifacts lockfiles.

Even though this Dockerfile assumes that the pip dependencies will be
fetched as wheels, some additional build dependencies were still needed
in order to install these dependencies in an offline context. They can
be found in the added requirements-build.txt file. Note that this file
was generated manually based on the build errors that were found.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>